backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
psycopg2-binary      # or mysqlclient if you use MySQL
groq
langgraph
langchain-core
langchain-community
